Added CropCircle

+++ src/main/scala/demo/commandsAndEvents.scala
@@ -1,14 +1,14 @@
 package demo
 
 import akka.actor.typed.ActorRef
-import demo.Membership.Command
+import demo.ReplicatedShardCoordinator.Command
 
 sealed trait ShardRegionCmd
 case class IdentifyShard(replyTo: ActorRef[Command]) extends ShardRegionCmd
 
 sealed trait DeviceCommand extends ShardRegionCmd {
-  def id: Int
+  def id: Long
   def replica: String
 }
 
-case class PingDevice(id: Int, replica: String) extends DeviceCommand
+case class PingDevice(id: Long, replica: String) extends DeviceCommand
